+++ src/tools/ToolsCata/vmap_tools/vmap4_extractor/gameobject_extract.cpp
@@ -104,8 +104,8 @@ void ExtractGameobjectModels()
 
         if (result)
         {
-            uint32 displayId = it->getUInt(0);
-            uint32 path_length = strlen(name);
+            uint32_t displayId = it->getUInt(0);
+            uint32_t path_length = static_cast<uint32>(strlen(name));
             fwrite(&displayId, sizeof(uint32), 1, model_list);
             fwrite(&path_length, sizeof(uint32), 1, model_list);
             fwrite(name, sizeof(char), path_length, model_list);
