+++ src/tools/ToolsTBC/vmap_tools/vmap4_extractor/wmo.cpp
@@ -41,7 +41,7 @@ WMORoot::WMORoot(std::string &filename)
 bool WMORoot::open()
 {
     MPQFile f(filename.c_str());
-    if(f.isEof ())
+    if (f.isEof())
     {
         printf("No such file.\n");
         return false;
@@ -119,7 +119,7 @@ bool WMORoot::open()
         */
         f.seek((int)nextpos);
     }
-    f.close ();
+    f.close();
     return true;
 }
 
@@ -149,7 +149,7 @@ WMOGroup::WMOGroup(const std::string &filename) :
 bool WMOGroup::open()
 {
     MPQFile f(filename.c_str());
-    if(f.isEof ())
+    if (f.isEof())
     {
         printf("No such file.\n");
         return false;
@@ -390,7 +390,7 @@ int WMOGroup::ConvertToVMAPGroupWmo(FILE *output, WMORoot *rootWMO, bool precise
         fwrite(VERT,4,3,output);
         for (uint32 i=0; i<nVertices; ++i)
             if(IndexRenum[i] >= 0)
-                check -= fwrite(MOVT+3*i, sizeof(float), 3, output);
+                check -= static_cast<int>(fwrite(MOVT+3*i, sizeof(float), 3, output));
 
         assert(check==0);
 
@@ -516,7 +516,7 @@ WMOInstance::WMOInstance(MPQFile& f, char const* WmoInstName, uint32 mapID, uint
 
     fseek(input, 8, SEEK_SET); // get the correct no of vertices
     int nVertices;
-    int count = fread(&nVertices, sizeof (int), 1, input);
+	int count = static_cast<int>(fread(&nVertices, sizeof(int), 1, input));
     fclose(input);
 
     if (count != 1 || nVertices == 0)
@@ -549,7 +549,7 @@ WMOInstance::WMOInstance(MPQFile& f, char const* WmoInstName, uint32 mapID, uint
     fwrite(&scale, sizeof(float), 1, pDirfile);
     fwrite(&pos2, sizeof(float), 3, pDirfile);
     fwrite(&pos3, sizeof(float), 3, pDirfile);
-    uint32 nlen=strlen(WmoInstName);
+    uint32_t nlen=static_cast<uint32_t>(strlen(WmoInstName));
     fwrite(&nlen, sizeof(uint32), 1, pDirfile);
     fwrite(WmoInstName, sizeof(char), nlen, pDirfile);
